✨ api-serverがbffで読み込まれているテスト用のroute
