--- apps/api-server/src/utils/config/incdx.ts
@@ -1,3 +1,0 @@
-export const apiServerConfig = {
-  baseUrl: process.env.API_SERVER_BASE_URL,
-};
